+++ test/cookbooks/filebeat_test/recipes/previous-ver.rb
@@ -1,0 +1,17 @@
+filebeat_install 'default' do
+  version '6.8.8'
+end
+
+filebeat_config 'default' do
+  config node['filebeat_test']['filebeat_config']
+end
+
+node['filebeat_test']['prospectors'].each do |p_name, p_config|
+  filebeat_prospector p_name do
+    config p_config
+  end
+end
+
+filebeat_service 'default' do
+  purge_prospectors_dir true
+end